requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
Sphinx
django-extensions==1.5.5
Werkzeug==0.10.4
django-test-plus==1.0.7
factory_boy==2.5.2

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.3.2

# improved REPL
ipdb==0.8.1

# Enables better email testing
maildump==0.5.1
